backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
chromadb==0.4.18
openai==1.3.7
pypdf2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
